backend/requirements.txt
Flask
pandas
scikit-learn
scikit-surprise
faker
